Database connection established, table for users created and methods for storing and fetching users created.
(!!) Installed the pg package using npm install pg

(!!) you need a postgresql database on your local computer and reestablish the link using the ports, names, users etc. of the pool. this is because we do not have a server to put the databse on, so it needs to be local.